Updates NpipePluginServer implementation
This commit updates the NpipePluginServer to use the Docker
provided Volume handler for servicing API requests. The Get and
List volume APIs will be fully functional on Windows with this
commit.

Caveat: Paths are case-insensitive on Windows, so Docker only
supports lower case volume names. To handle this, we use a
platform specific normalizeVolumeName func in the vmdk_driver.

* Updates NpipePluginServer impl to use Docker Volume handler.
* Updates NpipePluginServer unit test.
* Adds handling for case-insensitivity in volume names on Windows.
* Updates logging in the fs module.